+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.7.14 (2021-09-29)
+# 5.7.14 (2021-09-30)
 
 ### 🐣新特性
 * 【extra  】     修复HttpCookie设置cookies的方法，不符合RFC6265规范问题（issue#I4B70D@Gitee）
@@ -17,6 +17,7 @@
 * 【core   】     TreeBuilder增加部分根节点set方法（issue#1848@Github）
 * 【core   】     优化Base64.isBase64方法：减少一次多余的判断（pr#1860@Github）
 * 【core   】     优化Base64.isBase64方法：减少一次多余的判断（pr#1860@Github）
+* 【cache  】     优化FIFOCache未设置过期策略时，无需遍历判断过期对象（pr#425@Gitee）
 
 ### 🐞Bug修复
 * 【http   】     修复HttpCookie设置cookies的方法，不符合RFC6265规范问题（pr#418@Gitee）